% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{as.schema_name}
\alias{as.schema_name}
\title{Creates a schema.
Get topic name}
\usage{
as.schema_name(x, project = Sys.getenv("GCP_PROJECT"))
}
\arguments{
\item{x}{\code{character}, \code{Topic}}
}
\value{
\code{character}
}
\description{
Creates a schema.
Get topic name
}
\keyword{internal}
