% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{as.snapshot_name}
\alias{as.snapshot_name}
\title{Get snaphsot name}
\usage{
as.snapshot_name(x, project = Sys.getenv("GCP_PROJECT"))
}
\arguments{
\item{x}{\code{character}, \code{Topic}}
}
\value{
\code{character}
}
\description{
Get snaphsot name
}
\keyword{internal}
