% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{as.sub_name}
\alias{as.sub_name}
\title{Get topic name}
\usage{
as.sub_name(x, project = Sys.getenv("GCP_PROJECT"))
}
\arguments{
\item{x}{\code{character}, \code{Topic}}
}
\value{
\code{character}
}
\description{
Get topic name
}
\keyword{internal}
