% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{as.topic_name}
\alias{as.topic_name}
\title{Get topic name}
\usage{
as.topic_name(x, project = Sys.getenv("GCP_PROJECT"))
}
\arguments{
\item{x}{\code{character}, \code{Topic}}
}
\value{
(\code{character})
}
\description{
Get topic name
}
\keyword{internal}
