% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{DlqPolicy}
\alias{DlqPolicy}
\title{Builds a DqlPolicy object}
\usage{
DlqPolicy(dlq_topic, max_delivery_attempts)
}
\arguments{
\item{dlq_topic}{\code{character}, \code{Topic} Required, topic name or instance of a topic object}

\item{max_delivery_attempts}{\code{numeric}}
}
\value{
\code{DlqPolicy} object
}
\description{
Builds a DqlPolicy object
}
\seealso{
Other Object functions: 
\code{\link{DeadLetterPolicy}()},
\code{\link{ExpirationPolicy}()},
\code{\link{MessageStoragePolicy}()},
\code{\link{PubsubMessage}()},
\code{\link{PushConfig}()},
\code{\link{RetryPolicy}()},
\code{\link{SchemaSettings}()},
\code{\link{Schema}()},
\code{\link{Snapshot}()},
\code{\link{Subscription}()},
\code{\link{Topic}()}
}
\concept{Object functions}
