% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ExpirationPolicy}
\alias{ExpirationPolicy}
\title{Builds a ExpirationPolicy Object}
\usage{
ExpirationPolicy(ttl = NULL)
}
\arguments{
\item{ttl}{\code{numeric} Specifies the 'time-to-live' duration (in seconds, can be float) for an
associated resource}
}
\value{
\code{ExpirationPolicy} object
}
\description{
Builds a ExpirationPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A policy that specifies the conditions for resource expiration (i.e., automatic resource deletion).
}
\seealso{
Other Object functions: 
\code{\link{DeadLetterPolicy}()},
\code{\link{DlqPolicy}()},
\code{\link{MessageStoragePolicy}()},
\code{\link{PubsubMessage}()},
\code{\link{PushConfig}()},
\code{\link{RetryPolicy}()},
\code{\link{SchemaSettings}()},
\code{\link{Schema}()},
\code{\link{Snapshot}()},
\code{\link{Subscription}()},
\code{\link{Topic}()}
}
\concept{Object functions}
