% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlePubsubR.R
\docType{package}
\name{googlePubsubR}
\alias{googlePubsubR}
\title{R interface for Google Pub/Sub REST API}
\description{
This library offers easy to use R interface for Google Pub/Sub REST API. Functions are
prefixed with the object type they deal with (e.g \verb{topics_*}).
\subsection{Resources:}{
\itemize{
\item \href{https://cloud.google.com/pubsub/docs/reference/rest}{Here} for the official Google
documentation
\item library \href{https://andodet.github.io/googlePubsubR/}{documentation}
}
}
}
