% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{MessageStoragePolicy}
\alias{MessageStoragePolicy}
\title{Builds a MessageStoragePolicy object}
\usage{
MessageStoragePolicy(regions)
}
\arguments{
\item{regions}{\code{character} A list of IDs of GCP regions}
}
\value{
\code{MessageStoragePolicy} object
}
\description{
Builds a MessageStoragePolicy object
}
\seealso{
Other Object functions: 
\code{\link{DeadLetterPolicy}()},
\code{\link{DlqPolicy}()},
\code{\link{ExpirationPolicy}()},
\code{\link{PubsubMessage}()},
\code{\link{PushConfig}()},
\code{\link{RetryPolicy}()},
\code{\link{SchemaSettings}()},
\code{\link{Schema}()},
\code{\link{Snapshot}()},
\code{\link{Subscription}()},
\code{\link{Topic}()}
}
\concept{Object functions}
