% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_decode}
\alias{msg_decode}
\title{Decode Pub/Sub message}
\usage{
msg_decode(x)
}
\arguments{
\item{x}{A base64 encoded string}
}
\value{
A deserialized object
}
\description{
Converts a Pub/Sub message into an object
}
\examples{
\dontrun{
library(jsonlite)

pulled_msgs$receivedMessages$messages$data \%>\% 
  msg_decode() \%>\%
  fromJSON()
}

}
