% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_encode}
\alias{msg_encode}
\title{Encode Pub/Sub message}
\usage{
msg_encode(x)
}
\arguments{
\item{x}{A serializeable object}
}
\value{
\code{character} a base64 encoded string
}
\description{
Converts an object into a base64 string
}
\examples{
\dontrun{
library(jsonlite)

mtcars \%>\% 
  toJSON(auto_unbox = TRUE) \%>\%
  msg_encode() \%>\% 
  PubsubMessage()
}

}
