% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ps_project_get}
\alias{ps_project_get}
\title{Get GCP projectId}
\usage{
ps_project_get()
}
\value{
\code{character} A valid GCP projectId, defaults to \code{GCP_PROJECT} env var
}
\description{
Get GCP projectId
}
\seealso{
Other Auth functions: 
\code{\link{ps_project_set}()},
\code{\link{pubsub_auth}()}
}
\concept{Auth functions}
