% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{pubsub_auth}
\alias{pubsub_auth}
\title{Authenticate a Pub/Sub client}
\usage{
pubsub_auth(
  json_file = Sys.getenv("GCP_AUTH_FILE"),
  token = NULL,
  email = NULL
)
}
\arguments{
\item{json_file}{\code{character} Path of the JSON file containing credentials for a GCP service
account}

\item{token}{\code{character} An existing authentication token}

\item{email}{\code{character} The email to default authentication to}
}
\value{
None, called for side effects
}
\description{
Authenticate a Pub/Sub client
}
\seealso{
Other Auth functions: 
\code{\link{ps_project_get}()},
\code{\link{ps_project_set}()}
}
\concept{Auth functions}
