% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{PushConfig}
\alias{PushConfig}
\title{Builds a PushConfig Object}
\usage{
PushConfig(attributes = NULL, push_endpoint = NULL, oidcToken = NULL)
}
\arguments{
\item{attributes}{\code{list} The attributes object or list of objects}

\item{push_endpoint}{\code{character} A URL locating the endpoint to which messages should
be pushed}

\item{oidcToken}{\code{character} If specified, Pub/Sub will generate and attach an OIDC JWT
token as an \code{Authorization} header in the HTTP request for every pushed message}
}
\value{
\code{PushConfig} object
}
\description{
Builds a PushConfig Object
}
\seealso{
Other Object functions: 
\code{\link{DeadLetterPolicy}()},
\code{\link{DlqPolicy}()},
\code{\link{ExpirationPolicy}()},
\code{\link{MessageStoragePolicy}()},
\code{\link{PubsubMessage}()},
\code{\link{RetryPolicy}()},
\code{\link{SchemaSettings}()},
\code{\link{Schema}()},
\code{\link{Snapshot}()},
\code{\link{Subscription}()},
\code{\link{Topic}()}
}
\concept{Object functions}
