% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Schema}
\alias{Schema}
\title{Builds a Schema Object}
\usage{
Schema(type = NULL, definition = NULL, name = NULL)
}
\arguments{
\item{type}{\code{character} The type of the schema definition}

\item{definition}{\code{character} The definition of the schema}

\item{name}{\code{character} The schema name}
}
\value{
\code{Schema} object
}
\description{
Builds a Schema Object
}
\seealso{
Other Object functions: 
\code{\link{DeadLetterPolicy}()},
\code{\link{DlqPolicy}()},
\code{\link{ExpirationPolicy}()},
\code{\link{MessageStoragePolicy}()},
\code{\link{PubsubMessage}()},
\code{\link{PushConfig}()},
\code{\link{RetryPolicy}()},
\code{\link{SchemaSettings}()},
\code{\link{Snapshot}()},
\code{\link{Subscription}()},
\code{\link{Topic}()}
}
\concept{Object functions}
