% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas_create}
\alias{schemas_create}
\title{Creates a schema}
\usage{
schemas_create(
  name,
  type = c("AVRO", "PROTOCOL_BUFFER", "TYPE_UNSPECIFIED"),
  definition,
  project = ps_project_get()
)
}
\arguments{
\item{name}{\code{character}, \code{Schema} Required, schema name or instance of a schema object}

\item{type}{\code{character} Type of the schema definition}

\item{definition}{\code{character} Required, the definition of the schema}

\item{project}{\code{character} GCP project id}
}
\value{
a \code{Schema} object
}
\description{
Creates a schema
}
\seealso{
Other Schema functions: 
\code{\link{schemas_delete}()},
\code{\link{schemas_exists}()},
\code{\link{schemas_get}()},
\code{\link{schemas_list}()},
\code{\link{schemas_validate_message}()},
\code{\link{schemas_validate}()}
}
\concept{Schema functions}
