% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas_delete}
\alias{schemas_delete}
\title{Deletes a schema}
\usage{
schemas_delete(name, project = ps_project_get())
}
\arguments{
\item{name}{\code{character}, \code{Schema} Schema name or instance of a schema object}

\item{project}{\code{character} GCP project id}
}
\value{
None, called for side effects
}
\description{
Deletes a schema
}
\seealso{
Other Schema functions: 
\code{\link{schemas_create}()},
\code{\link{schemas_exists}()},
\code{\link{schemas_get}()},
\code{\link{schemas_list}()},
\code{\link{schemas_validate_message}()},
\code{\link{schemas_validate}()}
}
\concept{Schema functions}
