% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas_exists}
\alias{schemas_exists}
\title{Check if a schema exists}
\usage{
schemas_exists(schema)
}
\arguments{
\item{schema}{\code{character}, \code{Schema} Required, an instance of a \code{Schema} object}
}
\value{
\code{logical}
}
\description{
Check if a schema exists
}
