% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SchemaSettings}
\alias{SchemaSettings}
\title{SchemaSettings Object}
\usage{
SchemaSettings(encoding = NULL, schema = NULL)
}
\arguments{
\item{encoding}{\code{character} The encoding of messages validated against \code{schema}}

\item{schema}{\verb{Schema, character} Required, schema object or schema name}
}
\value{
\code{SchemaSettings} object
}
\description{
SchemaSettings Object
}
\seealso{
Other Object functions: 
\code{\link{DeadLetterPolicy}()},
\code{\link{DlqPolicy}()},
\code{\link{ExpirationPolicy}()},
\code{\link{MessageStoragePolicy}()},
\code{\link{PubsubMessage}()},
\code{\link{PushConfig}()},
\code{\link{RetryPolicy}()},
\code{\link{Schema}()},
\code{\link{Snapshot}()},
\code{\link{Subscription}()},
\code{\link{Topic}()}
}
\concept{Object functions}
