% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{snapshots_exists}
\alias{snapshots_exists}
\title{Check if a snapshot exists}
\usage{
snapshots_exists(snapshot)
}
\arguments{
\item{snapshot}{\code{character}, \code{Snapshot} Required, snapshot name or an instance of a \code{Snapshot} object}
}
\value{
\code{logical} TRUE if snapshot exists
}
\description{
Check if a snapshot exists
}
\seealso{
Other Snapshot functions: 
\code{\link{snapshots_create}()},
\code{\link{snapshots_delete}()},
\code{\link{snapshots_list}()},
\code{\link{snapshots_patch}()}
}
\concept{Snapshot functions}
