% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_ack}
\alias{subscriptions_ack}
\title{Acknowledges the messages}
\usage{
subscriptions_ack(ack_ids, subscription)
}
\arguments{
\item{ack_ids}{\code{character} A vector containing one or more message ackIDs}

\item{subscription}{\code{character}, \code{Subscription} Required, the subscription whose messages
are being aknowledged}
}
\value{
\code{logical}
}
\description{
The Pub/Sub system can remove the relevant messages from the subscription.
Acknowledging a message whose ack deadline has expired may succeed, but such a message
may be redelivered later. Acknowledging a message more than once will not result in an error.
}
\seealso{
Other AcknowledgeRequest functions: 
\code{\link{subscriptions_seek}()}
}
\concept{AcknowledgeRequest functions}
