% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_detach}
\alias{subscriptions_detach}
\title{Detaches a subscription from a topic.}
\usage{
subscriptions_detach(subscription)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, subscription name or instance of
a \code{Subscription} object}
}
\description{
Detaches a subscription from a topic.
}
