% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_exists}
\alias{subscriptions_exists}
\title{Check if a subscription exists}
\usage{
subscriptions_exists(subscription)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, subscription name or instance of
a \code{Subscription} object}
}
\value{
\code{logical}
}
\description{
Check if a subscription exists
}
