% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_get}
\alias{subscriptions_get}
\title{Gets the configuration details of a subscription.}
\usage{
subscriptions_get(subscription)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, subscription name or instance of
a \code{Subscription} object}
}
\value{
A \code{Subscription} object
}
\description{
Gets the configuration details of a subscription.
}
