% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_list}
\alias{subscriptions_list}
\title{Lists matching subscriptions.}
\usage{
subscriptions_list(
  project = Sys.getenv("GCP_PROJECT"),
  pageSize = NULL,
  pageToken = NULL
)
}
\arguments{
\item{project}{\code{character} Required, GCP project id}

\item{pageSize}{\code{numeric} Maximum number of subscriptions to return}

\item{pageToken}{\code{character} The value returned by the last \code{subscriptions_list};
indicates that this is a continuation of a prior \code{subscriptions_list} call}
}
\value{
\code{data.frame}
}
\description{
Lists matching subscriptions.
}
