% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_modify_ack_deadline}
\alias{subscriptions_modify_ack_deadline}
\title{Updates an existing subscription}
\usage{
subscriptions_modify_ack_deadline(subscription, ack_ids, ack_deadline)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} A subscription name or \code{Subscription} object}

\item{ack_ids}{\code{character} A vector containing ackIDs. They can be acquired using}

\item{ack_deadline}{\code{numeric} The new ack deadline (in seconds)}
}
\value{
\code{logical}
}
\description{
Certain properties of a subscription, such as its topic, are not modifiable.
}
\seealso{
Other UpdateSubscriptionRequest functions: 
\code{\link{subscriptions_patch}()}
}
\concept{UpdateSubscriptionRequest functions}
