% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_modify_pushconf}
\alias{subscriptions_modify_pushconf}
\title{Modify PushConfig for a subscription}
\usage{
subscriptions_modify_pushconf(subscription, push_config)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, a subscription name or a \code{Subscription}
object}

\item{push_config}{\code{PushConfig} New PushConfig object, can be built using \code{\link{PushConfig}}}
}
\value{
\code{logical}
}
\description{
Modify PushConfig for a subscription
}
