% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_pull}
\alias{subscriptions_pull}
\title{Pulls messages from the server.}
\usage{
subscriptions_pull(subscription, max_messages = 100)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, subscription where to pull
messages from}

\item{max_messages}{\code{numeric} Maximum number of messages to return}
}
\value{
A named \code{list} with pulled messages
}
\description{
Pulls messages from the server.
}
\seealso{
Other Subscription functions: 
\code{\link{subscriptions_ack}()},
\code{\link{subscriptions_create}()},
\code{\link{subscriptions_delete}()},
\code{\link{subscriptions_detach}()},
\code{\link{subscriptions_exists}()},
\code{\link{subscriptions_get}()},
\code{\link{subscriptions_list}()},
\code{\link{subscriptions_modify_ack_deadline}()},
\code{\link{subscriptions_modify_pushconf}()},
\code{\link{subscriptions_patch}()},
\code{\link{subscriptions_seek}()}
}
\concept{Subscription functions}
