% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_seek}
\alias{subscriptions_seek}
\title{Seek a subscription to a point in timex}
\usage{
subscriptions_seek(subscription, time = NULL, snapshot = NULL)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, a snapshot name or a \code{Snapshot}
object}

\item{time}{\code{character} A timestamp in RFC3339 UTC "Zulu" format}

\item{snapshot}{\code{character}, \code{Snapshot} A Snapshot name or a \code{Snapshot} object}
}
\value{
\code{logical}
}
\description{
A subscription can be seeked to a point in time or to a given snapshot.
}
\seealso{
Other AcknowledgeRequest functions: 
\code{\link{subscriptions_ack}()}
}
\concept{AcknowledgeRequest functions}
