% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_delete}
\alias{topics_delete}
\title{Deletes a pub/sub topic}
\usage{
topics_delete(topic)
}
\arguments{
\item{topic}{\code{character}, \code{Topic} Required, topic name or instance of a topic object}
}
\description{
Deletes a pub/sub topic
}
