% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_exists}
\alias{topics_exists}
\title{Check if a topic exists}
\usage{
topics_exists(topic, project = Sys.getenv("GCP_PROJECT"))
}
\arguments{
\item{topic}{\code{character}, \code{Topic} Required, topic name or instance of a topic object}

\item{project}{\code{character} Required, GCP project id}
}
\value{
\code{logical}
}
\description{
Check if a topic exists
}
