% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_get}
\alias{topics_get}
\title{Gets a topic configuration}
\usage{
topics_get(topic)
}
\arguments{
\item{topic}{\code{character}, \code{Topic} Required, instance of \code{Topic} or \code{character}}
}
\value{
\code{Topic} A \code{Topic} object
}
\description{
Gets a topic configuration
}
