% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_list_subscriptions}
\alias{topics_list_subscriptions}
\title{Lists the names of the attached subscriptions on this topic.}
\usage{
topics_list_subscriptions(topic, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{topic}{\code{Topic}, \code{character} Required, an instance of a \code{Topic} object or a string of the
topic name}

\item{pageToken}{The value returned by the last response; indicates that this is a continuation
of a prior \code{topics_list_subscriptions()} paged call, and that the system should return the next
page of data}

\item{pageSize}{Maximum number of subscription names to return}
}
\value{
A \code{character} vector
}
\description{
Lists the names of the attached subscriptions on this topic.
}
