% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{UpdateObjectRequest}
\alias{UpdateObjectRequest}
\title{Build update request for an object}
\usage{
UpdateObjectRequest(object = NULL)
}
\arguments{
\item{object}{An object for which preparing a patch request}
}
\value{
An \verb{Update*} object to be passed as a request body
}
\description{
Build update request for an object
}
\keyword{internal}
