% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ValidateMessageRequest}
\alias{ValidateMessageRequest}
\title{ValidateMessageRequest Object}
\usage{
ValidateMessageRequest(
  message = NULL,
  name = NULL,
  schema = NULL,
  encoding = NULL
)
}
\arguments{
\item{message}{Message to validate against the provided \code{schema_spec}}

\item{name}{Name of the schema against which to validate}

\item{schema}{Ad-hoc schema against which to validate}

\item{encoding}{The encoding expected for messages}
}
\value{
\code{ValidateMessageRequest} object
}
\description{
ValidateMessageRequest Object
}
\seealso{
Other Object functions: 
\code{\link{DeadLetterPolicy}()},
\code{\link{DlqPolicy}()},
\code{\link{ExpirationPolicy}()},
\code{\link{MessageStoragePolicy}()},
\code{\link{PubsubMessage}()},
\code{\link{PushConfig}()},
\code{\link{RetryPolicy}()},
\code{\link{SchemaSettings}()},
\code{\link{Schema}()},
\code{\link{Snapshot}()},
\code{\link{Subscription}()},
\code{\link{Topic}()}
}
\concept{Object functions}
\keyword{internal}
