% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ps_project_set}
\alias{ps_project_set}
\title{Set GCP projectId}
\usage{
ps_project_set(project_id)
}
\arguments{
\item{project_id}{\code{character} A valid GCP projectId}
}
\value{
\code{character} ProjectId string
}
\description{
Set GCP projectId
}
\examples{
\dontrun{
ps_project_set("my-new-project")
# Do whatever...
# Jump back on the default project
ps_project_set(Sys.getenv("GCP_PROJECT"))
}

}
\seealso{
Other Auth functions: 
\code{\link{ps_project_get}()},
\code{\link{pubsub_auth}()}
}
\concept{Auth functions}
