% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RetryPolicy}
\alias{RetryPolicy}
\title{Builds a retry policy object}
\usage{
RetryPolicy(min_backoff = 600, max_backoff = 600)
}
\arguments{
\item{min_backoff}{\code{numeric} The minimum delay between consecutive deliveries of a
given message}

\item{max_backoff}{\code{numeric} The maximum delay between consecutive deliveries of a
given message}
}
\value{
\code{RetryPolicy} object
}
\description{
More on this \href{https://cloud.google.com/pubsub/docs/reference/rest/v1/projects.subscriptions#RetryPolicy}{here}
}
\seealso{
Other Object functions: 
\code{\link{DeadLetterPolicy}()},
\code{\link{DlqPolicy}()},
\code{\link{ExpirationPolicy}()},
\code{\link{MessageStoragePolicy}()},
\code{\link{PubsubMessage}()},
\code{\link{PushConfig}()},
\code{\link{SchemaSettings}()},
\code{\link{Schema}()},
\code{\link{Snapshot}()},
\code{\link{Subscription}()},
\code{\link{Topic}()}
}
\concept{Object functions}
