% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas_validate}
\alias{schemas_validate}
\title{Validates a schema}
\usage{
schemas_validate(schema, project = Sys.getenv("GCP_PROJECT"))
}
\arguments{
\item{schema}{\code{Schema} Required, an instance of a \code{Schema} object}

\item{project}{\code{character} GCP project id}
}
\value{
\code{logical} TRUE if successfully validated
}
\description{
Validates a schema
}
\seealso{
Other Schema functions: 
\code{\link{schemas_create}()},
\code{\link{schemas_delete}()},
\code{\link{schemas_exists}()},
\code{\link{schemas_get}()},
\code{\link{schemas_list}()},
\code{\link{schemas_validate_message}()}
}
\concept{Schema functions}
