% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{snapshots_delete}
\alias{snapshots_delete}
\title{Removes an existing snapshot}
\usage{
snapshots_delete(snapshot)
}
\arguments{
\item{snapshot}{\code{Snapshot}, \code{character} Required, an instance of a \code{Snapshot} object or a
object or a subscription name}
}
\value{
None, called for side effects
}
\description{
Removes an existing snapshot
}
\seealso{
Other Snapshot functions: 
\code{\link{snapshots_create}()},
\code{\link{snapshots_exists}()},
\code{\link{snapshots_list}()},
\code{\link{snapshots_patch}()}
}
\concept{Snapshot functions}
