% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{snapshots_get}
\alias{snapshots_get}
\title{Gets the configuration details of a snapshot}
\usage{
snapshots_get(snapshot)
}
\arguments{
\item{snapshot}{\code{Snapshot}, \code{character} Required, an instance of a \code{Snapshot} object or a
snapshot name}
}
\value{
An instance of a \code{Snapshot} object
}
\description{
Gets the configuration details of a snapshot
}
