% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{snapshots_list}
\alias{snapshots_list}
\title{Lists the existing snapshots}
\usage{
snapshots_list(project = ps_project_get(), pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{project}{\code{character} a GCP project ID}

\item{pageSize}{\code{numeric} Maximum number of snapshots to return}

\item{pageToken}{\code{character} The value returned by the last \code{ListSnapshotsResponse};
indicates that this is a continuation of a prior \code{ListSnapshots} call, and that the
system should return the next page of data}
}
\value{
A \code{data.frame} containing all snapshots
}
\description{
Lists the existing snapshots
}
\seealso{
Other Snapshot functions: 
\code{\link{snapshots_create}()},
\code{\link{snapshots_delete}()},
\code{\link{snapshots_exists}()},
\code{\link{snapshots_patch}()}
}
\concept{Snapshot functions}
