% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_detach}
\alias{subscriptions_detach}
\title{Detaches a subscription from a topic.}
\usage{
subscriptions_detach(subscription)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, subscription name or instance of
a \code{Subscription} object}
}
\value{
\code{logical}, TRUE if successfully detached
}
\description{
Detaches a subscription from a topic.
}
\seealso{
Other Subscription functions: 
\code{\link{subscriptions_ack}()},
\code{\link{subscriptions_create}()},
\code{\link{subscriptions_delete}()},
\code{\link{subscriptions_exists}()},
\code{\link{subscriptions_get}()},
\code{\link{subscriptions_list}()},
\code{\link{subscriptions_modify_ack_deadline}()},
\code{\link{subscriptions_modify_pushconf}()},
\code{\link{subscriptions_patch}()},
\code{\link{subscriptions_pull}()},
\code{\link{subscriptions_seek}()}
}
\concept{Subscription functions}
