% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_exists}
\alias{subscriptions_exists}
\title{Check if a subscription exists}
\usage{
subscriptions_exists(subscription)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, subscription name or instance of
a \code{Subscription} object}
}
\value{
\code{logical} TRUE if the subscription exist
}
\description{
Check if a subscription exists
}
\seealso{
Other Subscription functions: 
\code{\link{subscriptions_ack}()},
\code{\link{subscriptions_create}()},
\code{\link{subscriptions_delete}()},
\code{\link{subscriptions_detach}()},
\code{\link{subscriptions_get}()},
\code{\link{subscriptions_list}()},
\code{\link{subscriptions_modify_ack_deadline}()},
\code{\link{subscriptions_modify_pushconf}()},
\code{\link{subscriptions_patch}()},
\code{\link{subscriptions_pull}()},
\code{\link{subscriptions_seek}()}
}
\concept{Subscription functions}
