% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_list}
\alias{subscriptions_list}
\title{List subscriptions}
\usage{
subscriptions_list(
  project = Sys.getenv("GCP_PROJECT"),
  pageSize = NULL,
  pageToken = NULL
)
}
\arguments{
\item{project}{\code{character} Required, GCP project id}

\item{pageSize}{\code{numeric} Maximum number of subscriptions to return}

\item{pageToken}{\code{character} The value returned by the last \code{subscriptions_list};
indicates that this is a continuation of a prior \code{subscriptions_list} call}
}
\value{
\code{list}
}
\description{
List subscriptions
}
\seealso{
Other Subscription functions: 
\code{\link{subscriptions_ack}()},
\code{\link{subscriptions_create}()},
\code{\link{subscriptions_delete}()},
\code{\link{subscriptions_detach}()},
\code{\link{subscriptions_exists}()},
\code{\link{subscriptions_get}()},
\code{\link{subscriptions_modify_ack_deadline}()},
\code{\link{subscriptions_modify_pushconf}()},
\code{\link{subscriptions_patch}()},
\code{\link{subscriptions_pull}()},
\code{\link{subscriptions_seek}()}
}
\concept{Subscription functions}
