% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_modify_ack_deadline}
\alias{subscriptions_modify_ack_deadline}
\title{Modify the ack deadline for a subscription}
\usage{
subscriptions_modify_ack_deadline(subscription, ack_ids, ack_deadline)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} A subscription name or \code{Subscription} object}

\item{ack_ids}{\code{character} A vector containing ackIDs. They can be acquired using}

\item{ack_deadline}{\code{numeric} The new ack deadline (in seconds)}
}
\value{
\code{logical} TRUE if successfully modified
}
\description{
This method is useful to indicate that more time is needed to process a message by the
subscriber, or to make the message available for redelivery if the processing was
interrupted.
}
\seealso{
Other Subscription functions: 
\code{\link{subscriptions_ack}()},
\code{\link{subscriptions_create}()},
\code{\link{subscriptions_delete}()},
\code{\link{subscriptions_detach}()},
\code{\link{subscriptions_exists}()},
\code{\link{subscriptions_get}()},
\code{\link{subscriptions_list}()},
\code{\link{subscriptions_modify_pushconf}()},
\code{\link{subscriptions_patch}()},
\code{\link{subscriptions_pull}()},
\code{\link{subscriptions_seek}()}
}
\concept{Subscription functions}
