% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{subscriptions_modify_pushconf}
\alias{subscriptions_modify_pushconf}
\title{Modify PushConfig for a subscription}
\usage{
subscriptions_modify_pushconf(subscription, push_config)
}
\arguments{
\item{subscription}{\code{character}, \code{Subscription} Required, a subscription name or a \code{Subscription}
object}

\item{push_config}{\code{PushConfig} New PushConfig object, can be built using \code{\link{PushConfig}}}
}
\value{
\code{logical}, TRUE if successfully modified
}
\description{
Modify PushConfig for a subscription
}
\seealso{
Other Subscription functions: 
\code{\link{subscriptions_ack}()},
\code{\link{subscriptions_create}()},
\code{\link{subscriptions_delete}()},
\code{\link{subscriptions_detach}()},
\code{\link{subscriptions_exists}()},
\code{\link{subscriptions_get}()},
\code{\link{subscriptions_list}()},
\code{\link{subscriptions_modify_ack_deadline}()},
\code{\link{subscriptions_patch}()},
\code{\link{subscriptions_pull}()},
\code{\link{subscriptions_seek}()}
}
\concept{Subscription functions}
