% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_delete}
\alias{topics_delete}
\title{Deletes a pub/sub topic}
\usage{
topics_delete(topic)
}
\arguments{
\item{topic}{\code{character}, \code{Topic} Required, topic name or instance of a \code{Topic} object}
}
\value{
None, called for side effects
}
\description{
Deletes a pub/sub topic
}
\seealso{
Other Topic functions: 
\code{\link{topics_create}()},
\code{\link{topics_exists}()},
\code{\link{topics_get}()},
\code{\link{topics_list_subscriptions}()},
\code{\link{topics_list}()},
\code{\link{topics_patch}()},
\code{\link{topics_publish}()}
}
\concept{Topic functions}
