% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_exists}
\alias{topics_exists}
\title{Check if a topic exists}
\usage{
topics_exists(topic, project = ps_project_get())
}
\arguments{
\item{topic}{\code{character}, \code{Topic} Required, topic name or instance of a topic object}

\item{project}{\code{character} GCP project id}
}
\value{
\code{logical}, TRUE if topic exists, FALSE otherwise
}
\description{
Check if a topic exists
}
\seealso{
Other Topic functions: 
\code{\link{topics_create}()},
\code{\link{topics_delete}()},
\code{\link{topics_get}()},
\code{\link{topics_list_subscriptions}()},
\code{\link{topics_list}()},
\code{\link{topics_patch}()},
\code{\link{topics_publish}()}
}
\concept{Topic functions}
