% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_get}
\alias{topics_get}
\title{Gets a topic configuration}
\usage{
topics_get(topic)
}
\arguments{
\item{topic}{\code{character}, \code{Topic} Required, topic name or instance of a \code{Topic}}
}
\value{
\code{Topic}, A \code{Topic} object
}
\description{
Gets a topic configuration
}
\seealso{
Other Topic functions: 
\code{\link{topics_create}()},
\code{\link{topics_delete}()},
\code{\link{topics_exists}()},
\code{\link{topics_list_subscriptions}()},
\code{\link{topics_list}()},
\code{\link{topics_patch}()},
\code{\link{topics_publish}()}
}
\concept{Topic functions}
