% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_list}
\alias{topics_list}
\title{Lists topics from project}
\usage{
topics_list(
  project = Sys.getenv("GCP_PROJECT"),
  pageSize = NULL,
  pageToken = NULL
)
}
\arguments{
\item{project}{\code{character} GCP project id}

\item{pageSize}{\code{numeric} Maximum number of topics to return}

\item{pageToken}{\code{character} The value returned by the last \code{ListTopicsResponse}; indicates
that this is a continuation of a prior \code{ListTopics} call, and that the system should return the
next page of data.}
}
\value{
A \code{list} of topics
}
\description{
Lists topics from project
}
\seealso{
Other Topic functions: 
\code{\link{topics_create}()},
\code{\link{topics_delete}()},
\code{\link{topics_exists}()},
\code{\link{topics_get}()},
\code{\link{topics_list_subscriptions}()},
\code{\link{topics_patch}()},
\code{\link{topics_publish}()}
}
\concept{Topic functions}
