% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_list_subscriptions}
\alias{topics_list_subscriptions}
\title{List attached subscriptions to a topic.}
\usage{
topics_list_subscriptions(topic, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{topic}{\code{Topic}, \code{character} Required, an instance of a \code{Topic} object or a topic name}

\item{pageToken}{\code{character} The value returned by the last response; indicates that this is a continuation
of a prior \code{topics_list_subscriptions()} paged call, and that the system should return the next
page of data}

\item{pageSize}{\code{numeric} Maximum number of subscription names to return}
}
\value{
A \code{character} vector
}
\description{
List attached subscriptions to a topic.
}
\seealso{
Other Topic functions: 
\code{\link{topics_create}()},
\code{\link{topics_delete}()},
\code{\link{topics_exists}()},
\code{\link{topics_get}()},
\code{\link{topics_list}()},
\code{\link{topics_patch}()},
\code{\link{topics_publish}()}
}
\concept{Topic functions}
